Adds dark mode option
