--- .storybook/preview.ts
@@ -1,17 +1,0 @@
-import type { Preview } from "@storybook/react";
-
-import './global.css';
-
-const preview: Preview = {
-  parameters: {
-    actions: { argTypesRegex: "^on[A-Z].*" },
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/i,
-      },
-    },
-  },
-};
-
-export default preview;
