+++ .storybook/preview.tsx
@@ -1,0 +1,50 @@
+import type { Preview } from "@storybook/react";
+import React from 'react';
+import { lightTheme, darkTheme } from '@backstage/theme';
+import { CssBaseline, ThemeProvider } from "@mui/material";
+
+export const withMuiTheme = (Story, context) => {
+  const theme = context.parameters.theme || context.globals.theme
+  const storyTheme = theme === 'dark' ? darkTheme : lightTheme
+  return (
+    <ThemeProvider theme={storyTheme}>
+      <CssBaseline />
+      <Story />
+    </ThemeProvider>
+  );
+};
+
+export const decorators = [withMuiTheme];
+
+const preview: Preview = {
+  parameters: {
+    actions: { argTypesRegex: "^on[A-Z].*" },
+    controls: {
+      matchers: {
+        color: /(background|color)$/i,
+        date: /Date$/i,
+      },
+    },
+  },
+};
+
+export const globalTypes = {
+  theme: {
+    name: 'Theme',
+    description: 'Global theme for components',
+    defaultValue: 'light',
+    toolbar: {
+      // The icon for the toolbar item
+      icon: 'circlehollow',
+      // Array of options
+      items: [
+        { value: 'light', icon: 'circlehollow', title: 'light' },
+        { value: 'dark', icon: 'circle', title: 'dark' },
+      ],
+      // Property that specifies if the name of the item will be displayed
+      showName: true,
+    },
+  },
+}
+
+export default preview;
